aider: Fix TypeError by correctly passing and defining windowPos object in isWithinSlidingWindow and findFarthestPosition functions.
# Aider chat conversation:

USER: Please address the following error: TypeError: undefined is not an object (evaluating 'windowPos.x')

ASSISTANT: The error `TypeError: undefined is not an object (evaluating 'windowPos.x')` suggests that when the `windowPos` object is being accessed, it is `undefined`. This can happen if the `windowPos` object is not passed correctly to the function or if it is not defined in the scope where it is being used.

To address this error, we need to ensure that the `windowPos` object is correctly passed to the `isWithinSlidingWindow` function and that it is defined within the scope of the function that is trying to access it.

Let's correct the `isWithinSlidingWindow` function call by ensuring that the `windowPosition` object is passed to it. We will also make sure that the `windowPosition` object is defined in the scope of the `findFarthestPosition` function.

Here is the change that needs to be made in the `2048slide.js` file:

```javascript
2048slide.js
<<<<<<< SEARCH
function isWithinSlidingWindow(x, y, windowPos, windowSize) {
=======
function isWithinSlidingWindow(x, y) {
    let windowPos = windowPosition;
    let windowSize = WINDOW_SIZE;
>>>>>>> REPLACE
```

And for the `findFarthestPosition` function, we need to ensure that `windowPosition` is passed correctly:

```javascript
2048slide.js
<<<<<<< SEARCH
let {newX, newY, merged} = findFarthestPosition(x, y, vector, windowPosition, WINDOW_SIZE);
=======
let {newX, newY, merged} = findFarthestPosition(x, y, vector);
>>>>>>> REPLACE
```

```javascript
2048slide.js
<<<<<<< SEARCH
function findFarthestPosition(x, y, vector, windowPos, windowSize) {
=======
function findFarthestPosition(x, y, vector) {
    let windowPos = windowPosition;
    let windowSize = WINDOW_SIZE;
>>>>>>> REPLACE
```

Please apply these changes to the `2048slide.js` file to resolve the error. If there are any other instances where `windowPos` is used incorrectly, similar changes will need to be made.