aider: Added color support for 4096, 8192, and five more powers of two tiles.
# Aider chat conversation:

USER: add color support for 4096 8192 and 5 more powers of two tiles.

ASSISTANT: To add color support for 4096, 8192, and five more powers of two tiles, we need to modify the `2048slide.css` file. We will add new CSS classes for each of these new tile values.

Here are the *SEARCH/REPLACE* blocks for the `2048slide.css` file:

2048slide.css
```css
<<<<<<< SEARCH
.tile.value-2048 { background-color: #edc22e; color: #f9f6f2; }
=======
.tile.value-2048 { background-color: #edc22e; color: #f9f6f2; }
.tile.value-4096 { background-color: #3c3a32; color: #f9f6f2; }
.tile.value-8192 { background-color: #0f4c81; color: #f9f6f2; }
.tile.value-16384 { background-color: #800080; color: #f9f6f2; }
.tile.value-32768 { background-color: #008080; color: #f9f6f2; }
.tile.value-65536 { background-color: #808000; color: #f9f6f2; }
.tile.value-131072 { background-color: #008000; color: #f9f6f2; }
.tile.value-262144 { background-color: #800000; color: #f9f6f2; }
>>>>>>> REPLACE
```

Please note that the colors chosen for the new tile values are arbitrary. You may want to adjust the colors to fit the design of your game.