aider: Adjusted colors for tiles starting from 4096 and heading into the green color scheme.
# Aider chat conversation:

USER: please adjust those arbitraty colors to build from 4096 and head into the green color scheme.

ASSISTANT: To adjust the colors for the tiles starting from 4096 and heading into the green color scheme, we will modify the `2048slide.css` file. We will update the CSS classes for each of the new tile values to have a gradient of green colors.

Here are the *SEARCH/REPLACE* blocks for the `2048slide.css` file:

2048slide.css
```css
<<<<<<< SEARCH
.tile.value-4096 { background-color: #3c3a32; color: #f9f6f2; }
.tile.value-8192 { background-color: #0f4c81; color: #f9f6f2; }
.tile.value-16384 { background-color: #800080; color: #f9f6f2; }
.tile.value-32768 { background-color: #008080; color: #f9f6f2; }
.tile.value-65536 { background-color: #808000; color: #f9f6f2; }
.tile.value-131072 { background-color: #008000; color: #f9f6f2; }
.tile.value-262144 { background-color: #800000; color: #f9f6f2; }
=======
.tile.value-4096 { background-color: #31a354; color: #f9f6f2; } /* Darker green */
.tile.value-8192 { background-color: #63c583; color: #f9f6f2; } /* Medium green */
.tile.value-16384 { background-color: #85d79b; color: #f9f6f2; } /* Lighter green */
.tile.value-32768 { background-color: #a8e8b4; color: #f9f6f2; } /* Light green */
.tile.value-65536 { background-color: #cbf9cc; color: #f9f6f2; } /* Very light green */
.tile.value-131072 { background-color: #efffe5; color: #f9f6f2; } /* Almost white green */
.tile.value-262144 { background-color: #ffffff; color: #f9f6f2; } /* White to indicate max value */
>>>>>>> REPLACE
```

Please adjust the colors as needed to fit the design of your game. The colors provided are a gradient from a darker green to white, indicating increasing tile values.